added method to retrieve stages

+++ XFifaFantasy/X-FIFA Fantasy Cup/Models/Player.cs
@@ -17,7 +17,7 @@ namespace X_FIFA_Fantasy_Cup.Models
         public int player_height { get; set; }
         public int player_weight { get; set; }
         public string player_team { get; set; }
-        public int player_price { get; set; }
+        public float player_price { get; set; }
         public int player_active { get; set; }
         public string player_photo { get; set; }
